+++ Sources/VCore/Services and Managers/Network Client/Network Request Factory/NetworkRequestFactory.Headers.swift
@@ -38,7 +38,7 @@ extension NetworkRequestFactory {
             for (key, value) in json {
                 guard let value = value else { continue }
                 
-                guard let description: String = .init(unwrappedDescribing: value) else { throw NetworkError.invalidHeaders }
+                guard let description: String = .init(unwrappedDescribing: value) else { throw NetworkClientError.invalidHeaders }
                 result.updateValue(description, forKey: key)
             }