+++ Sources/VCore/Views and ViewControllers/CapsuleUIView.swift
@@ -1,0 +1,26 @@
+//
+//  CapsuleUIView.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 26.07.22.
+//
+
+#if canImport(UIKit) && !os(watchOS)
+
+import UIKit
+
+// MARK: - Capsule View
+/// `UIView` that rounds corners to capsule.
+open class CapsuleUIView: UIView {
+    // MARK: Lifecycle
+    open override func layoutSubviews() {
+        super.layoutSubviews()
+        
+        // `clipsToBounds` and `layer.maskedCorners` are not necessary.
+        // Plus, they conflict with shadows.
+        
+        layer.cornerRadius = frame.height / 2
+    }
+}
+
+#endif