+++ CHANGELOG.md
@@ -15,6 +15,25 @@ XCode Templates
 
 - Project now includes templates for `SwiftUI`
 
+### [3.18.0(53)](https://github.com/VakhoKontridze/VCore/releases/tag/3.18.0) — *2022 07 26*
+
+View and ViewControllers
+
+- `CapsuleUIView` is added
+- `CapsuleUIImageView` is added
+
+Network Client
+
+- `NetworkError` is renamed to `NetworkClientError`
+
+Architectural Pattern Helpers
+
+- `func pop(count: Int, animated: Bool)` is added to `StandardNavigable`
+
+Extensions
+
+- `UIStackView.removeArrangedSubviewsFromSuperview()` is added
+
 ### [3.17.0(52)](https://github.com/VakhoKontridze/VCore/releases/tag/3.17.0) — *2022 07 23*
 
 Services and Managers